Fix "Rainbow" button and slider bug.
All buttons are now working as expected.
Previously upon clicking the reset button, slider value would reset on the backened side but not on the frontend side for the users to see. This issue is now solved by setting "slider.value" to default in the event handling function of the resetBtn.